+++ SonsGameManager/Tests/SoundTests.cs
@@ -1,5 +1,8 @@
-﻿using RedLoader;
+﻿using FMOD;
+using RedLoader;
 using SonsSdk;
+using TheForest;
+using TheForest.Utils;
 using UnityEngine;
 using Color = System.Drawing.Color;
 
@@ -13,19 +16,48 @@ public class SoundTests
     public static void Init()
     {
         GlobalEvents.OnUpdate.Subscribe(OnUpdate);
+        SdkEvents.OnGameStart.Subscribe(OnGameStart);
         
         var path = @"C:\Users\Julian\Downloads\DDOI.mp3";
-        SoundTools.RegisterSound("ddoi", path, true);
+        var sound = SoundTools.RegisterSound("ddoi", path, true);
+        sound?.set3DMinMaxDistance(1, 5);
         Log("Registered sound!");
     }
+
+    private static void OnGameStart()
+    {
+        DebugConsole.RegisterCommand("fmodist", (Il2CppSystem.Func<string, bool>)SetDistance, DebugConsole.Instance);
+    }
+
+    private static bool SetDistance(string args)
+    {
+        var parts = args.Split(' ').Select(float.Parse).ToArray();
+        var min = parts[0];
+        var max = parts[1];
+
+        Player.ChannelDistance = (min, max);
+        RLog.Debug($"Distances {min} {max}");
+        return true;
+    }
+    
 
     private static void OnUpdate()
     {
         if (Input.GetKeyDown(KeyCode.G))
         {
-            var go = ActorTools.GetRobby().gameObject;
-            SoundTools.BindSound(go, "ddoi").Play();
-            Log("Playing sound on Robby!");
+            var go = GetObjectToAttachTo();
+            Player = SoundTools.BindSound(go, "ddoi");
+            Player.Play();
+            Log("Playing sound");
         }
     }
+
+    private static GameObject GetObjectToAttachTo()
+    {
+        var go = new GameObject("Sound Player");
+        go.transform.position = LocalPlayer.Transform.position;
+        return go;
+    }
+
+    public static SoundPlayer Player { get; set; }
 }