+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Console position can now be set in the config
 - (Experimental!) Console position can automatically be saved on game exit (doesn't handle dpi when console is on different monitor with different resolution)
 - Added additional SdkEvents like OnItemPickup, OnItemConsumed, OnItemCrafted and OnArmorEquipped
+- Added `--savegame <saveid>` console argument to load a specific savegame on start